stricter use of types, fixing lots of compiler warnings [Michael D. Adams]
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1525 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/CSeries/cscluts_sdl.cpp
@@ -47,7 +47,7 @@ RGBColor system_colors[NUM_SYSTEM_COLORS] =
 void build_color_table(color_table *table, LoadedResource &clut)
 {
 	// Open stream to CLUT resource
-	SDL_RWops *p = SDL_RWFromMem(clut.GetPointer(), clut.GetLength());
+	SDL_RWops *p = SDL_RWFromMem(clut.GetPointer(), (int)clut.GetLength());
 	assert(p);
 
 	// Check number of colors
