+++ Source_Files/RenderMain/AnimatedTextures.cpp
@@ -50,11 +50,11 @@ class AnimTxtr
 	
 	// Phases: which frame, and which tick in a frame
 	// These must be in ranges (0 to NumFrames - 1) and (0 to abs(NumTicks) - 1)
-	short FramePhase, TickPhase;
+	size_t FramePhase, TickPhase;
 
 public:
 	// How many frames
-	int GetNumFrames() {return FrameList.size();}
+	size_t GetNumFrames() {return FrameList.size();}
 		
 	// Frame ID (writable)
 	short& GetFrame(int Index) {return FrameList[Index];}
@@ -112,11 +112,11 @@ void AnimTxtr::Load(vector<short>& _FrameList)
 bool AnimTxtr::Translate(short& Frame)
 {
 	// Sanity check
-	int NumFrames = FrameList.size();
+	size_t NumFrames = FrameList.size();
 	if (NumFrames <= 0) return false;
 
 	// Find the index in the loop; default: first one.
-	short FrameIndx = 0;
+	size_t FrameIndx = 0;
 	if (Select >= 0)
 	{
 		if (Frame != Select) return false;
@@ -124,7 +124,7 @@ bool AnimTxtr::Translate(short& Frame)
 	else
 	{
 		bool FrameFound = false;
-		for (int f=0; f<NumFrames; f++)
+		for (size_t f=0; f<NumFrames; f++)
 			if (FrameList[f] == Frame)
 			{
 				FrameFound = true;
@@ -167,7 +167,7 @@ void AnimTxtr::SetTiming(short _NumTicks, short _FramePhase, short _TickPhase)
 		FramePhase += TickPhaseFrames;
 	}
 	
-	int NumFrames = FrameList.size();
+	size_t NumFrames = FrameList.size();
 	if (NumFrames > 0)
 	{
 		FramePhase = FramePhase % NumFrames;
@@ -179,13 +179,13 @@ void AnimTxtr::SetTiming(short _NumTicks, short _FramePhase, short _TickPhase)
 void AnimTxtr::Update()
 {
 	// Be careful to wrap around in the appropriate direction
-	int NumFrames = FrameList.size();
+	size_t NumFrames = FrameList.size();
 	if (NumTicks > 0)
 	{
-		if (++TickPhase >= NumTicks)
+		if (++TickPhase >= (size_t)NumTicks)
 		{
 			TickPhase = 0;
-			if (++FramePhase >= NumFrames)
+			if (++FramePhase >= (size_t)NumFrames)
 				FramePhase = 0;
 		}
 	}
